+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## X.Y.Z (Unreleased)
-* Add new change notes here
+FEATURES:
+* **New Resource:** sumologic_azure_metrics_source (GH-710)
 
 ## 3.0.0 (December 09, 2024)
 **REMOVALS:**